kio_travelup_be/requirements.txt
fastapi
uvicorn[standard]
python-jose[crypotgraphy]
passlib[bcrypt]
SQLAlchemy